Sahayak-/sahayak_plus/requirements.txt
Flask
requests
python-dotenv
google-generativeai
Pillow
reportlab

# Eye tracking dependencies
opencv-python
mediapipe
pyautogui
websockets
asyncio
